v2.1
Oh well, we suppose people lie sometimes :P

This updates was necessary as we forgot to complete the requirements.

- Added a review system, now users can add ratings from 1 to 5 stars.
- Fixed bugs related to the cart page when items are out-of-stock (speaking of which, yes we've added out-of-stock stylings).
- Added a way to reset password, new much-needed button on the login page.
- We've added EMAILS! Now users will receive emails regarding their order (when placing order, when shipping, and when delivered).
- Added the recommendation system in the home page as well.
- Adjusted CSS of recommendation containers.
- Added KU logo in the home page (temporary; we plan on making our own).
- Generated a sample set of users, orders, and ratings (for demo reasons).
- Added more analytics; admins can now see the best and worst rated products.

We're now confident that at least 99% of the project requirements are done now, future updates will either be bug fixes, or minor updates (such as logo, or increasing user-friendliness) until the project is submitted.

It was our pleasure to ride this journey. As a team, we've learned a lot from building websites, especially using React and Node js, which assisted us in building proper front and back end (full stack) development.

Co-Authored-By: mshorafa45 <167345453+mshorafa45@users.noreply.github.com>
Co-Authored-By: Conan <151556080+wqixx@users.noreply.github.com>